requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# PDF processing
pypdf2==3.0.1
langchain-text-splitters==0.0.1

# Vector database and embeddings
faiss-cpu==1.7.4
langchain-community==0.0.13
langchain==0.0.335
sentence-transformers==2.2.2

# LLM integration
openai==1.3.5

# Utilities
tqdm==4.66.1
numpy==1.26.0
pandas==2.1.1